Get the code to compile again.  After Ted's latest round, my build was failing with:
  error: cast from ‘Rosegarden::Segment*’ to ‘int’ loses precision

This seemed like a no brainer to me.  Just remove the cast, and carry on.


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@12784 cf6a83e7-f017-4305-b0b3-678bde24f230

+++ src/gui/seqmanager/CompositionMapper.cpp
@@ -98,7 +98,7 @@ CompositionMapper::segmentDeleted(Segment *segment)
 {
     // Converting segment to a number to avoid dereferencing a pointer to
     // deleted memory.
-    SEQMAN_DEBUG << "CompositionMapper::segmentDeleted(" << (int)segment << ")\n";
+    SEQMAN_DEBUG << "CompositionMapper::segmentDeleted(" << segment << ")\n";
 
     // !!! WARNING !!!
     // The segment pointer that is coming in to this routine has already
